chore: move entity loader utils into their own object
