+++ packages/entity/src/EntitySecondaryCacheLoader.ts
@@ -85,7 +85,7 @@ export default abstract class EntitySecondaryCacheLoader<
 
     // convert value to and from array to reuse complex code
     const entitiesMap = await this.entityLoader
-      .withAuthorizationResults()
+      .utils()
       .constructAndAuthorizeEntitiesAsync(
         mapMap(loadParamsToFieldObjects, (fieldObject) => (fieldObject ? [fieldObject] : [])),
       );
